Function to calculate strain invariants
MVT_strain_invariants - calculate inveriants of square symmetric matrix

Return I_E, II_E and III_E for a strain or strain rate tensor. I_E is the
trace and gives the volume change, II_E is a measure of strain magnitude
and III_E is the determinant. These do not change if the tensor is
rotated.

Usage:
    [I_E, II_E, III_E] = MVT_strain_invariants(E)
        E must be a 3*3 symmetric real matrix.